refactor(chat): lift SSE bootstrap into chat-sse-transport helper
The SSE branch was ~290 LoC of ReadableStream orchestration inline in
the route handler: encoder + send helper + heartbeat lifecycle + start
function body + truncation/empty-reply/persist/notify sequence. With
the legacy worker pipeline removed, this was the last meaningful
extraction left in routes/chat.ts.

Move it to src/routes/chat-sse-transport.ts as buildChatSseStream.
The route now calls:

  const stream = buildChatSseStream(params);
  return new Response(stream, { headers: STREAM_HEADERS, status: 200 });

The replyPersisted flag is shared with the outer route handler (it's
read in the post-error cleanup at chat.ts:1037), so the helper takes a
markReplyPersisted callback that the route flips when the row
transitions to COMPLETE.

routes/chat.ts: 1048 -> 794 LoC. The remainder is queue-mode dispatch,
in-process JSON path, status + retry endpoints. Drop two newly-unused
imports (notifyChatReplyCompleted, createChatSseEventHandlers) — both
now live in the transport helper.
